Fixes combustion turbine power grid (#19973)
### Introduction
The power grid of the new combustion turbine is behaving very weird. For
example, the TEG outputs 300kW of power, but inspecting the wire with a
multitool shows very random things like 12mW. The SMES shows a different
value as well.

### Fix
I noticed that a lot of cabling has been used with the d1,d2 properties
set, which the wiki states to not use. I replaced all cabling starting
from the TEG up to the main grid with cabling without the d1,d2
properties. This fixed all power values in the power network. The TEG,
cabling and SMES now show the exact same value of power available in the
network.

Additionally i added a powernet sensor to the turbine grid, as it was
missing.

_Edit: Grammar_