IAC Rescue Ship Bugfixes (#19977)
Started as me fixing the airlock and docking markers and ended with me
finding various little bugs that I missed when I remapped the ship a
while back.

- Properly creates subtypes for the airlock/shuttle/docking markers
specific to the IAC ship instead of just having them be variable edited
versions of the main type.
- Fixed some incorrectly named airlocks.
- Adjusted some floor decals.
- Tweaked some access requirements.
- Returns the uranium-powered portable generator I originally intended
for on account of #19978.