Fixes SCC & TCAF flight helmets (#19984)
The parent type for flight helmets defines unique head sprites for
Tajara and Unathi, intended for variations with a longer visor to
accommodate snouts. These aren't sprited for the SCC or TCAF variants,
so they just turned out invisible ingame. This sets the definition to
null for both of those until someone decides to sprite species
variations for them.

Moves the TCAF flight helmet to pilot_helmet.dm for neatness.

Resolves https://github.com/Aurorastation/Aurora.3/issues/19968