Re-adds stealth ling dna sting (#19986)
Adds the DNA extraction sting which currently exists in the code to the
changeling ability store, enabling the ability in-game.

From what I can tell, the reason the ability is currently disabled is
because it originally allowed changelings to gain more ability points
without any risk, which is no longer the case as of the merging of
https://github.com/Aurorastation/Aurora.3/pull/17140, which made it so
that changelings only get a set amount of ability points and cannot gain
more.

DNA extraction sting in its current form does not allow changelings to
respec, that still requires changelings to absorb someone.

---------

Signed-off-by: Tag114 <118570749+Tag114@users.noreply.github.com>