Small QoL & balance tweaks for some ghostroles (#19998)
Implements a few small tweaks to the Lone Spacer, Tramp Freighter, and
OE Ship that I've been sitting on for a while.

- Lone Spacer: opens up a few more clothing options, adds a suit cooler
for IPCs, adds dylovene, adds an Orion Express package receptacle, adds
some credits.
- Tramp Freighter: added OE packages, toned down the loot a little so
we're less likely to see an entire freighter crew parading in military
hardsuits, added shutters to the windows.
- OE Ship: removed a stray grate, added dynamic space under catwalks.