Melee Toy, Sprites  - Dino Grabber (#19900)
Adds a dino grabber toy to the game. It counts as a weapon, and uses
modified whip code to have a chance of disarming people. Its also now in
the warehouse spawn pool as a rare item, with the weight of 1.75. It
should only show up every 2-3 rounds.
Also adds an object sprite and in hand sprites.

---------

Signed-off-by: courierbravo <73250577+courierbravo@users.noreply.github.com>
Co-authored-by: Fluffy <65877598+FluffyGhoster@users.noreply.github.com>
Co-authored-by: Hazelrat <ric.whe.don@gmail.com>