Fixed Atmospheric technician spawn count (#20071)
### Description of the bug
The game is configured to allow three Atmos techs to spawn at
roundstart, there are also three spawn markers in the engineering
breakroom, but at round start only two out of three readied ATs spawn,
one gets thrown back into the lobby.

### Changes implemented
In the job code configuration of the Atmospheric technician the number
of spawn_positions was raised from two to the expected three.

### Testing
I am not aware on how the change can be tested locally. If it is
testable somehow, let me know.

_Edit: ~~lobby~~ breakroom_