Modular Armour Tweaks (#20062)
Changes the slowdown of modular armour:

- The plate accessory slowdown has been changed from 1 to 0.4
- The leg guard accessory slowdown has been changed from 0 to 0.3
- the arm guard accessory slowdown has been changed from 0 to 0.3

When wearing a full set of armour, your slowdown should be equal to 1
which is the current slowdown, taking off the leg and arm guards will
now allow you to move faster at the cost of coverage. This allows better
customisation and gives more incentive to play around with the modular
side of the modular armour system. Note that this slowdown does not
apply to light armour like the standard corporate armour and similar.