Makes vaurca surgery consistent (#20078)
It seems rather silly to have a 15% chance of making the situation worse
when operating on vaurca, when there is no equivalent for humans, etc.

Yes, this is a salt PR after the random chance just killed a Ta by
**deleting her phoron tank.**

I could see an argument for keeping it in since incision management
systems have no failure rate, but I don't see those that often,
especially in low pop. I guess it made the round more interesting but..