Makes the available ghostrole list display properly (#20109)
The pregame() proc was shuffled around in init order after the ghostrole
list was implemented into it, so it currently procs before the
ghostroles the list should reference have spawned in. This shifts all of
the 'cosmetic' elements of pregame over to a new welcome proc which is
called much later, so it properly reads out the available ghostroles.
The welcome message should also be a little more visible to users now
that it isn't so early in initialisation.

Current:

![image](https://github.com/user-attachments/assets/99aef9df-86a8-4f24-806c-4fe46e12ad2d)

With this PR:

![image](https://github.com/user-attachments/assets/c6f4b536-257f-4923-ab56-aa75f4e53484)