Some hostile mob code AI cleanup (#20116)
Fixed some situations where a null was being set as target.
Repathed the dreaded target_mob variable to last_found_target, and from
mob to atom, as it was not being used only for mobs.
Added procs to set and unset the last_found_target and setup the signals
relative to clear it, and validations.
Thanks to the above, removed a viewscan that was needed to delete
references when mobs die from their attacker.
General QoL of code, DMDocs.