SCCV Quark - new xenoarch/science shuttle (#20127)
feedback thread:
-
https://forums.aurorastation.org/topic/21275-sccv-quark-new-xenoarch-shuttle-feedback-thread/#comment-178259

changes:
- rscadd: "Adds SCCV Quark, new shuttle for xenoarch/science, on
Horizon."
- rscadd: "Small remap of the aux hangar on Horizon, adding some
expedition supplies there."
 
reasoning and explanation:
- The idea is that xenoarchs/science use this if they want to go on a
solo adventure for whatever reason (and die to random simplemobs),
without stealing the intrepid (that could be used to go to the odyssey
scenario site).
- Xenoarchs would still be allowed to use the intrepid, if it's not used
for odyssey or whatever and they want to organize a bigger expedition
with more crew (or at least I hope so, I don't make the rules).
- It's a small shuttle, but it has more than enough seating and some
cargo space. I think, at least, it is fine for a xenoarch or two and a
few passengers.
- We need a xenoarch/science shuttle like yesterday, and no one's really
shown any interest in mapping it in the past few months, so I made this.
- If anyone has a different idea and wants to map their own shuttle,
it's fine ofc. This is just my PR, anyone is free to open their own, and
the best shuttle will be merged



![image](https://github.com/user-attachments/assets/269f6369-c95a-4e78-8305-0525df30058b)

![image](https://github.com/user-attachments/assets/507fca8c-ea38-4a7b-b268-8b62acc7f543)

![image](https://github.com/user-attachments/assets/fc7857a6-2a4a-414a-a49b-3610a123c812)

![image](https://github.com/user-attachments/assets/984cfcc1-756b-4963-b6fe-b01912326eb6)

![image](https://github.com/user-attachments/assets/ca803886-925a-4449-a343-4c4f3acdbeba)

(explanation on top of description)

---------

Co-authored-by: DreamySkrell <>