Jaws of life for atmos (#20138)
### Introduction
This PR adds a hydraulic rescue tool (jaws of life) to the atmos locker
room. The goal is to provide Atmospheric technicians the means to enter
areas that they are needed in without knowing the airlock wiring or
having access to insulated gloves. This places them in a better position
for their emergency response duties. This also provides the opportunity
for cross-department work towards paramedics and maybe security.

The idea raised from this suggestion/thread: [Reinforce Engineering's
Division of
Labor](https://forums.aurorastation.org/topic/21285-reinforce-engineerings-division-of-labor/)

### What has been changed?
- Added a jaws of life to the table in the atmos locker room.
- Added the functionality to force open powered airlocks that are not
broken or bolted. Without this functionality the tool wouldn't serve its
intended purpose.
- Added a couple of messages when using the tool for different actions
for fluff.