Fixed ladder access (#20150)
### Introduction
With the port of SSthrowing from TG (see #19421) an additional check got
added to `/atom/movable/Move`. The check
```
if(!newloc || newloc == loc)
    return
```
prevents the move method from an unnecessary movement.
The method is also used by ladders to move the player to the ladder
before climbing it - With the new additional check it was no longer
possible to grab ladders if you are on the same tile, because the move
method would run the location check and return, skipping the movement,
which is correct, but results in failing to climb the ladder after all
("You fail to reach the ladder.").

This PR fixes this small but very annoying issue.

### What changed?

To not further modify the move method an additional check has been added
to the ladder-climb code to skip the move method when the player is
already on the same tile.