Hydraulic rescue tool refactoring (#20162)
### Introduction
The recently to atmos added hydraulic rescue tool is a robotic tool and
has none of the common item/tool properties. This PR refactors the tool
to a non-robotic version and fixes a bug in the recently added
functionality (#20138 - Airlocks can be opened instantly by moving
during the action duration).

### What changed?

- Added `/obj/item/crowbar/hydraulic_rescue_tool` to tools.dm
- Removed robotic version under AI
(`/obj/item/crowbar/robotic/jawsoflife`)
- Updated all occurrences (Sprite label due to updating the sprite
filename)
- Fix action skip bug in airlock.dm when using the tool (See
[5cc808b](https://github.com/Aurorastation/Aurora.3/pull/20162/commits/5cc808ba53787a2bf44be476e9503860e5440f9c))