Offsite scattershot changes & fixes (#20172)
Touches on the Sol Bunker, Shady Asteroid, and Crashed Elyran offsites
to bring them a little closer to standard. Alters balance of the Sol
Bunker by giving it a regular airlock that can be cycled. Extends the
docking hangar of the Shady Asteroid so that it should(?) support most
possible shuttles. Gives the Sol Bunker the old Solarian flag subtype -
since it spawns in the Coalition, I assume this is meant to be mid or
pre-IW.