Bodybag and buckling layer fixes (#20174)
This PR is mostly about fixing body/cryo bag layering issues. 

Currently live and annoying to try and treat around:

![image](https://github.com/user-attachments/assets/b867af9a-8ab1-42cc-8503-bb252a3a86c7)


Currently, bags are are below roller beds when first deployed. If
buckled and unbuckled, they are then above, because of the + 0.1 on
buckling.

If buckled on a roller bed, dragged, and then unbuckled, they are then
above player mobs, because of the `layer + 1` on roller bed `Move()`
proc, which also doesn't get reset. I removed this +1 on Move() because
it was older than the +0.1 and seems to serve the same purpose.

The +0.1 did not have a method of reverting to the original buckled
atom's layer, so I added a var and check to handle that.

I also changed bullet casing's layer to `BELOW_TABLE_LAYER` with the
intent of not having them on top of stasis bags and roller beds, because
it causes issues as someone trying to treat wounded after a firefight.

Changing bullet casing's layer does have a consequence of putting
bullets underneath, for example, the soil in the public garden. Putting
the soil on `BELOW_TABLE_LAYER` would fix this, but with the consequence
of having the bushes be on top, making clicking the soil difficult for
farming purposes. Adding bushes to that layer causes consequences with
side window layering (like on deck three outside the public lounge.)

For this PR I only adjusted the bullet casing layer. I'm sure it has
other unintended layering interactions. Let me know if it would be
better to leave them where they are or on some other layer.

Current changes in PR: Bullets under soil/roller beds/stasis bags. 

![image](https://github.com/user-attachments/assets/b56b768f-b9d6-4aca-ad2a-cfc30fe6e66a)


(Example of why making bullets on top of soil is difficult) Screenshot
shows both bullets and the soil being `BELOW_TABLE_LAYER` with bushes on
top, as a reference. PR Currently has bullets underneath soil.

![image](https://github.com/user-attachments/assets/8cbf4b6d-8023-4138-86f1-f4cac106f647)