Update the armory roles to account for the removal of the crew armory. (#20168)
This updates the armory roles to account for the removal of the crew
armory in #19728. In particular, these roles no longer have the role:

- Chief Engineer
- Chief Medical Officer
- Research Director

The rest of the roles on the list can access the security armory, or can
at least provide the access to the armory via an ID change.

The Operations Manager remains on the list. While they do not have
direct access to the armory, their department is allowed to distribute
guns, and this is an intended function of their role as stated in
#19728.

---------

Signed-off-by: Cody Brittain <1779662+Generalcamo@users.noreply.github.com>
Co-authored-by: Cody Brittain <cbrittain10@live.com>