Basic First Aid Kit Changes (#20187)
Changes the contents of first aid kits.

* Antidexafen pills replaced with inaprovaline pills
* Emergency autoinjector added
* Health Analyzer removed

This PR when coupled with #20194 _negligibly affects the number of
health analyzers actually available to the crew and offships._ Any
non-medical personnel that previously relied on access to health
analyzers will in almost every case still have access to them. The
removal is being done to keep the number of analyzers on the map
similar, while increasing the accessibility to stabilization supplies.
If this change is not made, with #20194 there will be over a dozen more
analyzers floating around the ship, which is unwanted by the developer
team. This **_only affects the basic white square medkits._**

If you find any spots where a shuttle or ghost spawn no longer has
access to any form of scanning, let me know in the comments of #20194 so
I can add the analyzer or an alternative there.

---------

Signed-off-by: OolongCow <61329630+OolongCow@users.noreply.github.com>
Co-authored-by: Wowzewow (Wezzy) <42310821+alsoandanswer@users.noreply.github.com>