Gargoyle voidsuit resprite (#20200)
This implements a resprite of the regular Gargoyle voidsuit, the SFA
variant, and the SSMD variant, full credit to shqvelheaded on discord
for the sprites. This also adds a unique variant for the SPLF, which I
intend to use for https://github.com/Aurorastation/Aurora.3/pull/20197.

Not all of the Gargoyle variants have been resprited, but most of the
ones that are accessible ingame short of using the uplink, and are
extant factions in the setting, have been.


![image](https://github.com/user-attachments/assets/9e5b7058-738f-49ae-9997-333dcf84668c)