Adds a Solarian army peacekeeper beret to the loadout. (#20129)
https://github.com/user-attachments/assets/b2ee7517-ca3a-4915-918b-319b518e1cb9

I have Schwann's blessing to add this

From the extended description;
"While a common sight in the Alliance's more rebellious or devastated
worlds such as Mars, the Army peacekeeper beret became a potent symbol
of Sol's triumph in the civil war. Nowhere else in the Spur are the blue
berets more omnipresent than the former capital of the Restoration Front
- Lycoris."

Available in the headgear tab- I've put it as an independent item since
the beret selection is for departmental / non factional berets, if
desired, I can move it under there.

---------

Signed-off-by: Matt Atlas <mattiathebest2000@hotmail.it>
Co-authored-by: Matt Atlas <mattiathebest2000@hotmail.it>