Adds Assunzionii voidsuit modkits to loadout (#20224)
This creates two modkits for the Assunzionii 'Rook' voidsuit, one for
the human variant and one for the IPC variant, available in the loadout
to any character with an Assunzionii origin. It's only applicable to
research voidsuits.

IPCs have access to both variants so human-shaped synthetics can still
get a suit they can use.

In current code, the Rook is a military voidsuit descended from the
Coalition Vulture. As this allows you to transfer it from research
voidsuits, it has been shifted to be a mechanically identical subtype of
the standard research suit. Anything in the description alluding to its
military uses or its anomaly resistance have been removed to reflect
this.

This should be a relatively clean change, the Rook isn't currently used
substantially in the uplink or for any ghostroles. My thought is that,
if someone ever does make an Assunzione military offship, they could
resprite this suit for a new armoured variety. The implementation should
otherwise resemble the implementation of the Himean modkits.