Slowdown removal from Captain's Voidsuit. (#20234)
Removes the slowdown from the Captain's voidsuit to be on par with the
rest of the voidsuits.

No armor values have been changed. The voidsuit is behind a WL for the
most part, and has been for a long time. Captain's should be able to
flee stuff while wearing their voidsuit without walking in slow motion.

---------

Signed-off-by: Butterrobber202 <buttersgt@gmail.com>