Added a hardsuit for Solarian Special Forces (#19846)
I had this sprite lying around for a potential rework of our Operative
outfits. This might still happen, but with Odyssey coming sooner or
later, I felt it was better to get this implemented now for usage by
storytellers and actors.

![image](https://github.com/user-attachments/assets/fdd391d5-ba3a-4a6c-a4e5-b6f6d0032ddf)

This is a specialized variant of the Vampire hardsuit, for Solarian
Special Forces. The commando variant (the only one implemented) has
deathsquad-tier stats. This should not be used lightly, but as a
definitely and obviously hostile alternative to the Asset Protection
teams currently implemented.

---------

Co-authored-by: Cody Brittain <cbrittain10@live.com>