Adds an *alarm audible emote (with sound) for IPCs (#20265)
Does exactly what the title says. The specific sound used is the warning
buzzer sound that plays when a delivery chute or disposal unit is about
to release an object.

Full disclosure, I did this because I specifically wanted this. I wanted
to be able to do this.
But, here's my retroactive justification for why it's *needed*:

IPCs do not have a sound emote to express _distress_. Sure, *buzz
exists, but that's more an expression of frustration or anger (the sound
is even called "buzz sigh"). Distress, though? Being caught in a corner
as the mercs with a dominian-army gimmick start blazing gunfire down the
corridor? Sure, you could use *scream, but that's very _human_. Sounding
an alarm? Now that's exactly what a soulless machine would do.