SFA spawning tweaks (#20253)
This removes the SFA ghostrole from the CRZ and reduces its spawn weight
in the remaining sectors from 1 to 0.5. The SFA is meant to be a badly
waning gaggle of remnants, so they should be spawning less frequently,
and the new SPLF ghostrole fills the niche of a piratical Sol ship in
the CRZ.

Cleared by Schwann, so this should be fine on the lore side.