Various offship QoL changes (#20290)
See changelog, mostly a variety of very mundane QoL changes to a few
offships - hydroponics had the wrong vendor on the SPLF ship, adjusting
misaligned firedoors, accidentally miscoloured windows on the Volvalaad
ship, etc.

This replaces the fake object lattices in the freebooter ship with
actual lattices. I don't believe it's traditionally possible to overlay
lattices on plating ingame, but it resolves a lot of the jank with the
fake objects.

The freebooter now spawns with a ballistic carbine, .45 machine pistol,
and 7.62 bolt action rifle in the armoury every round, plus one
additional random handgun. It also spawns with 2 armored voidsuits every
round, the other 3 suits remaining random. 50/50 randomised to static is
intended to make the ship a little stronger, so it's impossible you roll
poor RNG and end up with 5 non-combat voidsuits and a bunch of pocket
rifles in the armory.