Fix devices shocking players repeatedly  (#20325)
A fix for https://github.com/Aurorastation/Aurora.3/issues/18219

Thanks to mattatlas for figuring out the actual cause -
`/datum/wires/suit_storage_unit/interactable` proc was called every tick
when TGUI window for that element was opened, and it had code that
checked for `isElectrified()`, leading to repeated shocks.

Removal should resolve the issue, actual .shock() calls are happening
when appropriate - in `/datum/wires/<MACHINE>/on_pulse(wire)` when
interacting with wiring, or in `<MACHINE>.dm` itself when processing
interactions.
 
Could not do full testing due to a lack of experience with the game and
tooling, but no issues were encountered with what was done.

---------

Co-authored-by: color <color@WINEWEASEL>