Changes options for Orepit citizenship and updates important origin i… (#20313)
**REQUIRES SYNTH LORE TEAM APPROVAL!!!** If accepted, all added blurbs
are placeholders that can be altered as the synth lore team sees fit.

The reasoning regarding this PR is that I think the current in-game
information around Orepit and the available listed options for
citizenship are too confusing for anyone who may be interested in the
origins; humans can either choose Coalition of Colonies or no
citizenship, despite implied statelessness for humans being confusing
for anyone who might read their card or records and Orepit not being
integrated with the broader Coalition. This makes Orepit a separate and
unique citizenship in a similar vein to the format for the Free Tajaran
Council or Elyran Non-citizen Persons. The included important
information blurbs are for clarity of lore developments regarding the
Open Doors memorandum and megacorporate hiring standards for Trinarists.

I understand that I was told by NM that plans will be in the works for
Orepit after the Golden Deep arc and hope this isn't seen as
overstepping or a stopgap measure.


![image](https://github.com/user-attachments/assets/3c48b59f-9019-4477-9c12-492d14acf1a0)
For humans

![image](https://github.com/user-attachments/assets/306cb756-4807-4623-827e-fbcacfca1fa0)
For IPCs

![image](https://github.com/user-attachments/assets/6c6c35fa-ff82-4052-b7f8-1b89ffca414a)

Co-authored-by: Gabby <59633984+feartheblackout@users.noreply.github.com>