Mix Pipe Connection to Bluespace Drive (#20380)
Adds a pipe connection from Atmospherics to the storage room behind the
Bluespace Drive. It's my understanding the intended mode of operating
the BSDrive was for Atmos Techs to lay down a pipe all the way to
Atmospherics, but it's an extremely time consuming and tedious process
to cross such a massive distance (maybe not point-to-point, but in
practice there's a great deal of twisting turns and doorways to go
through) with a pipeline, especially every single round. It's either a
huge timesink nobody wants to do for a "just in case" scenario, or it
takes too long to be viable if a jump is urgently needed.

Admittedly, I think a mix chamber in the BSDrive itself would be better,
but I don't want to go against the spirit of the intent.


![StrongDMM_hc2ZQf7iiW](https://github.com/user-attachments/assets/7a820482-69a1-4428-8d1d-1a0c6406a2d8)

![StrongDMM_Z3oLgG32NU](https://github.com/user-attachments/assets/f5bd60e6-148a-44df-8cb6-615efdaa2e90)

Personally I feel this is a nice compromise. It cuts past the hardest
and most tedious part of laying pipe (the maintenance area between
atmos/custodial/grav generator are extremely congested for pipe
traffic), but doesn't do the whole job for you. It lets you fill
canisters in the storage room to haul back into the BSDrive, or if you
want a direct line-in, it's close enough now that it won't be a 20
minute job.

The Bluespace Drive is awesome, and I'd love to see it used more.