Removes Access to Leave the Operating Theatres in Medical (#20412)
Adds one unrestricted door access effect to each operating theatre door
to allow everyone to leave.

It is not uncommon for Interns, Patients, and other crew to get stuck in
the ORs and have to beg over medical or common for someone to let them
out. This would fix that issue. Similar to how there is unrestricted
access on doors leaving the GTR.

Of course, those without access to surgery would still not be able to
enter the ORs.

---------

Signed-off-by: Acetrea <138949778+Acetrea@users.noreply.github.com>
Co-authored-by: hazelrat <83198434+hazelrat@users.noreply.github.com>