Reduces wheelchair loadout cost to 2 points (#20480)
Title. 4 loadout points is crazy for an item meant for roleplay value
and doesn't disincentivize anyone who would use the item for nefarious
purposes anyway.

More wheelchair changes to come whenever I figure out the code better,
including the ability to retain the name, description and color when
putting it down or picking it back up.

---------

Co-authored-by: Gabby <59633984+feartheblackout@users.noreply.github.com>