Lore adjustments to Orepit origins (#20525)
Done with permission from @NobleRow and the synthetic lore team!
Reflects the outcome of the BitByte news article [found
here.](https://forums.aurorastation.org/topic/15858-bitbyte/page/4/#comment-179689)

For humans: "All Orepitters abroad in the Open Doors memorandum would
not work in jobs not requiring an education, and cannot work for
Hephaestus Industries because of its practice of not hiring Trinarists,
nor directly for the Stellar Corporate Conglomerate due to its distrust
of the Church; Orepitters who travel abroad independently, either
because they are not Trinarists or not participants in the memorandum,
are not held to this. Human Orepitters born in the Twenty Parishes
should select the Native Orepitter accent, while humans born in
Providence or the Marches should select the Providence accent."

For IPCs: "All Orepitters abroad in the Open Doors memorandum would not
work in jobs not requiring an education, and cannot work for Hephaestus
Industries because of its practice of not hiring Trinarists, nor
directly for the Stellar Corporate Conglomerate due to its distrust of
the Church; Orepitters who travel abroad independently, either because
they are not Trinarists or not participants in the memorandum, are not
held to this."

---------

Co-authored-by: Gabby <59633984+feartheblackout@users.noreply.github.com>