Replaces a redundant camera in the Kitchen with a light fixture. (#20508)
This PR removes one of the cameras from the Kitchen (it is exactly two
tiles away from another camera that points in the same direction and
provides coverage of the same area) and puts a light fixture in its
place. This should hopefully brighten up the awkwardly dim spot around
the oven and freezer doors.

I can't add an image from the pull request function in VSC so I'm going
to add an image of the changes in a comment