No Orepit diplomatic aides (#20567)
Title. Characters with Orepit citizenship aren't meant to be either
Consular Officers or Diplomatic Aides, but the latter was showing up as
an option in the join menu so this should rectify that. Orepitters might
still able to be Diplomatic Aides for the Coalition depending on how
lore feels, but this doesn't touch that. Should count as a bugfix unless
maintainers decide it isn't.

---------

Co-authored-by: Gabby <59633984+feartheblackout@users.noreply.github.com>